+++ packages/dev/core/src/Misc/index.ts
@@ -2,7 +2,7 @@
 export * from "./andOrNotEvaluator";
 export * from "./assetsManager";
 export * from "./basis";
-export * from "./basisuEncoder";
+export * from "./basisEncoder";
 export * from "./dds";
 export * from "./decorators";
 export * from "./deferred";
